% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/getFileNames.R
\name{getFileNames}
\alias{getFileNames}
\title{Files currently available for download}
\usage{
getFileNames()
}
\value{
A named list of available single-season Retrosheet event and
game-log zip files, and schedule text files. These file names are
not intended to be passed to \code{getRetrosheet}, but is simply a
fast way to determine if the desired data is available.
}
\description{
A convenience function, returning the base file names of the
available downloads for the \code{year} and \code{type} arguments
in \code{getRetrosheet}.
}
\examples{
\donttest{
getFileNames()
}
}
