% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/getParkIDs.R
\name{getParkIDs}
\alias{getParkIDs}
\title{A data frame of ballpark IDs}
\usage{
getParkIDs()
}
\description{
This function returns a two-column data frame of ballpark IDs
along with current stadium name
}
\examples{
\donttest{
getParkIDs()
}

}
