% Generated by roxygen2 (4.1.0): do not edit by hand
% Please edit documentation in R/getPartialGamelog.R
\name{getPartialGamelog}
\alias{getPartialGamelog}
\title{Partial parser for game-log files}
\usage{
getPartialGamelog(year, glFields = getOption("retrosheet.glFields"),
  date = NULL, ...)
}
\arguments{
\item{year}{A single four-digit year.}

\item{glFields}{A subset of \code{getOption("retrosheet.glFields")}}

\item{date}{One of either NULL (the default), or a four-digit
character string identifying the date 'mmdd'}

\item{...}{further arguments passed to \code{\link[utils]{download.file}}}
}
\value{
A data table with dimensions
\code{length(date)} x \code{length(glFields)}
}
\description{
Choose the columns and rows
}
\details{

}
\examples{
## Get Homerun and RBI info for August 25, 2012
f <- grep("HR|RBI", getOption("retrosheet.glFields"), value = TRUE)
getPartialGamelog(2012, glFields = f, date = "0825")
}

