% Generated by roxygen2 (4.1.0): do not edit by hand
% Please edit documentation in R/getRetrosheet.R
\name{getRetrosheet}
\alias{getRetrosheet}
\alias{getRetrosheetTeamData}
\title{Import Retrosheet data as a structured R object}
\usage{
getRetrosheet(type, year, team, stringsAsFactors = FALSE, ...)

getRetrosheetTeamData(year, ...)
}
\arguments{
\item{type}{character.  This argument can take on either of "game" for
game-logs, "play" for play-by-play (a.k.a. event) data, "roster" for
team rosters, or "schedule" for the game schedule for the given year.}

\item{year}{integer. A valid four-digit year.}

\item{team}{character. Only to be used if \code{type = "play"}.
A single valid team ID for the given year. For available team IDs for the
given year call \code{getRetrosheetTeamData(year)}.  The available teams
are in the "TeamID" column.}

\item{stringsAsFactors}{logical. The \code{stringsAsFactors} argument as
used in \code{\link[base]{data.frame}}. Currently applicable to types "game" and "schedule".}

\item{...}{further arguments passed to \code{\link[utils]{download.file}}.}
}
\value{
The following return values are possible for the given \code{type}
\itemize{
\item \code{game} - a data frame of gamelog data for the given year
\item \code{play} - a list, each element of which is a single game's play-by-play
data for the given team and year.  Each list element is also a list, containing
the play-by-play data split into individual matrices.
\item \code{roster} - a named list, each element containing the roster
for the named team for the given year, as a data frame.
\item \code{schedule} - a data frame containing the game schedule for the given year
}
}
\description{
This function downloads and parses data from
\url{http://www.retrosheet.org} for the gamelog, play-by-play
(a.k.a. event), roster, and schedule files.
}
\details{

}
\examples{
\dontrun{

## gamelog data for the 2012 season
getRetrosheet("game", 2012)
## play-by-play data for the San Francisco Giants for the 2012 season
getRetrosheet("play", 2012, "SFN")
## a data frame of team information for the 2012 season
getRetrosheetTeamData(2012)
}
}

