% Generated by roxygen2 (4.1.0): do not edit by hand
% Please edit documentation in R/getFileNames.R
\name{getTeamData}
\alias{getTeamData}
\title{Retrieve a data frame of team information}
\usage{
getTeamData(year, ...)
}
\arguments{
\item{year}{integer. A valid four-digit year.}

\item{...}{further arguments passed to \code{\link[utils]{download.file}}.}
}
\value{
a data frame of team info for the given year
}
\description{
This function is for convenience, and returns a data frame of
team information, including team IDs, for the \code{team} argument in \code{getRetrosheet()}.
}

