% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/parse_event.R
\name{parse_event}
\alias{parse_event}
\title{NOT IMPLEMENTED Parse event strings from retrosheet play-by-play data
#'}
\usage{
parse_event(play)
}
\arguments{
\item{play}{A play object from \code{getRetrosheet(type = "play")} or \code{get_retrosheet(type = "play")}}
}
\value{
A list of intepreted details from the event strings#'
}
\description{
Parses the strings describing plays returned from the
#' \code{play} element of of \code{getRetrosheet("play", year, team)}.
}
\examples{
\donttest{
plays <- getRetrosheet("play", 2012, "SFN")[[1]]$play
parse_event(play)
}

}
