% Generated by roxygen2 (4.1.0): do not edit by hand
% Please edit documentation in R/getFileNames.R
\name{getFileNames}
\alias{getFileNames}
\title{Retrieve a list of current available downloads}
\usage{
getFileNames()
}
\value{
a named list of available Retrosheet zip and text files
}
\description{
This function is for convenience, and returns the file names
of available downloads for the \code{year} and \code{type} arguments
in \code{getRetrosheet()}
}
\details{

}

